[hyprland] [bash] lock with matrix script rework

+++ .config/hypr/lock_with_matrix.sh
@@ -1,13 +1,9 @@
 #!/bin/bash
 
-run_matrix="kitty --start-as=fullscreen -e cmatrix -C green -u 10 -b"
+run_matrix="kitty --start-as=fullscreen -e cmatrix -u 10 -b -s -C cyan"
 
-# if ! [ -z (pgrep hyprlock)]; then
-# screens=$(hyprctl -j monitors | jq length)
-# for ((i = -1; i < $screens; i++)); do
-# hyprctl dispatch focusmonitor $i
-eval $run_matrix
-hyprlock
-kill -9 $(pgrep -f "$run_matrix") # pkill cmatrix
-# done
-# fi
+if ! (pgrep -x cmatrix >/dev/null); then
+    eval $run_matrix
+    hyprlock
+    kill -9 $(pgrep -f "$run_matrix")
+fi
